+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - context: `Context::recv()` and `Context::try_recv()` panics if called again after returning `None`.
 - logger: use `_location` and `_module` instead of `@location` and `@module`.
 - logger: remove the cargo prefix from locations.
+- telemeter: hide rendered metrics in dumps.
 
 ## [0.1.27] - 2021-09-27
 ### Fixed